+++ packages/localize/examples/runtime/src/locales/zh_CN.ts
@@ -8,6 +8,6 @@
     /* eslint-disable @typescript-eslint/no-explicit-any */
 
     export const templates = {
-      'greeting': html`你好, <b>世界!</b>`,
+      '0h3c44aff2d5f5ef6b': html`你好, <b>世界!</b>`,
     };