+++ packages/localize/examples/transform/src/x-greeter.ts
@@ -3,7 +3,7 @@ import {msg} from '@lit/localize';
 
 export class XGreeter extends LitElement {
   render() {
-    return html`<p>${msg(html`Hello <b>World</b>!`, {id: 'greeting'})}</p>`;
+    return html`<p>${msg(html`Hello <b>World</b>!`)}</p>`;
   }
 }
 customElements.define('x-greeter', XGreeter);