+++ packages/localize/testdata/xliff/input/foo.ts
@@ -25,3 +25,9 @@ msg(
 );
 
 msg(html`Hello <b><!-- comment -->World!</b>`, {id: 'comment'});
+
+// Auto IDs
+msg(`Hello World!`);
+msg((name) => `Hello ${name}!`, {args: ['Friend']});
+msg(html`Hello <b>World</b>!`);
+msg((name) => html`Hello <b>${name}</b>!`, {args: ['Friend']});